+++ packages/vendure-plugin-invoices/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.1.0 (2024-08-30)
+
+- Exporting credit invoices via accounting strategies now have their own method interface
+- Don't allow accounting export when the order totals changed, to prevent mismatch between accounting export and invoice
+- Added Due Date to Xero exports, which is needed for invoice approval
+
 # 4.0.3 (2024-08-27)
 
 - Try to find Xero contact by name first, then by email address